refactor: excise more dead Python code
