--- vg_control/auth/__init__.py
@@ -1,1 +1,0 @@
-# Auth module initialization
