+++ vg_control/data/uploader.py
@@ -2,14 +2,12 @@ import os
 from typing import List, Optional
 from datetime import datetime
 import requests
-import urllib3
 import subprocess
 import shlex
 from tqdm import tqdm
 import time
 
 from ..constants import API_BASE_URL
-from ..metadata import get_hwid
 
 
 def _get_upload_url(
@@ -39,26 +37,6 @@ def _get_upload_url(
     response.raise_for_status()
     data = response.json()
     return data.get("url") or data.get("upload_url") or data["uploadUrl"]
-
-
-def _upload_archive(
-    api_key: str,
-    archive_path: str,
-    tags: Optional[List[str]] = None,
-    base_url: str = API_BASE_URL,
-) -> None:
-    """Upload an archive to the storage bucket via a pre-signed URL."""
-
-    upload_url = get_upload_url(
-        api_key,
-        archive_path,
-        tags=tags,
-        base_url=base_url,
-    )
-
-    with open(archive_path, "rb") as f:
-        put_resp = requests.put(upload_url, data=f, timeout=60, verify=False)
-        put_resp.raise_for_status()
 
 
 def get_upload_url(
@@ -304,3 +282,19 @@ def upload_archive(
 
         if return_code != 0:
             raise Exception(f"Upload failed with return code {return_code}")
+
+
+def get_hwid():
+    try:
+        with open("/sys/class/dmi/id/product_uuid", "r") as f:
+            hardware_id = f.read().strip()
+    except:
+        try:
+            # Fallback for Windows
+            import subprocess
+
+            output = subprocess.check_output("wmic csproduct get uuid").decode()
+            hardware_id = output.split("\n")[1].strip()
+        except:
+            hardware_id = None
+    return hardware_id
