--- vg_control/security/__init__.py
@@ -1,1 +1,0 @@
-# Security module initialization
